RedisClient: Move to lib/redis module
Code Climate started complaining after the previous commit during the
analysis of #174 of the lib/redis-client.js file exceeding 250 lines—and
still complained about duplicate code. This is the first step towards
extracting separate modules from the file, keeping them all contained
within the lib/redis package. Then a subsequent refactoring will address
the "duplicate" code warnings again.
