+++ index.js
@@ -11,7 +11,7 @@ var express = require('express')
 var app = express()
 var redis = require('redis')
 var LinkDb = require('./lib/link-db')
-var RedisClient = require('./lib/redis-client')
+var RedisClient = require('./lib/redis')
 var redisClientOptions = {}
 var session = require('express-session')
 var RedisStore = require('connect-redis')(session)
