+++ lib/redis/index.js
