+++ tests/server/redis-client-test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var RedisClient = require('../../lib/redis-client')
+var RedisClient = require('../../lib/redis')
 var helpers = require('../helpers')
 var redis = require('redis')
 var sinon = require('sinon')
